Split object command implementations
